move app protect dos funtions to their own file
